KeyDown.py
If you press escape while in the game window when you go to the console window, 
It'll show a bunch of w's - just press Escape or delete them if you need to quit [or press the X button like normal human beings],
Otherwise, if you press Enter, it'll just resume. If you click the console window and press Escape, you won't have that issue.